Add files via upload
This happens to be my best project of 2022 because it took me so much time (about 5 weeks) to start and complete it. I'm so happy that I am able to complete it the way I want despite all the challenges and I learnt a whole lot of new lessons and experience. Some of the inspirations about the contents came from Burger King and I'm grateful to them.